Move slow running tests into integration tests
